Move ApiResponse back to the api.py
