+++ fakturoid/model_api.py
@@ -3,33 +3,8 @@ import requests
 import re
 
 if TYPE_CHECKING:
-    from fakturoid.api import Fakturoid
+    from fakturoid.api import Fakturoid, APIResponse
     from fakturoid.models import Model
-
-
-LINK_HEADER_PATTERN = re.compile(r'page=(\d+)[^>]*>; rel="last"')
-
-def extract_page_link(header):
-    m = LINK_HEADER_PATTERN.search(header)
-    if m:
-        return int(m.group(1))
-    return None
-
-
-class APIResponse:
-    _requests_response: requests.Response
-
-    def __init__(self, requests_response: requests.Response):
-        self._requests_response = requests_response
-
-    def from_json(self):
-        return self._requests_response.json()
-
-    def page_count(self):
-        if 'link' in self._requests_response.headers:
-            return extract_page_link(self._requests_response.headers['link'])
-        else:
-            return None
 
 
 class ModelApi:
@@ -49,8 +24,8 @@ class ModelApi:
             raise ValueError("object wit unassigned id")
         return value.id
 
-    def from_response(self, api_response: APIResponse):
+    def from_response(self, api_response: 'APIResponse'):
         return self.model_type(**api_response.from_json())
 
-    def from_list_response(self, api_response: APIResponse):
+    def from_list_response(self, api_response: 'APIResponse'):
         return [self.model_type(**item) for item in api_response.from_json()]
